requirements.txt
# core
numpy<1.24
scipy
gym>=0.18.3,<=0.21.0
sapien>=2.2.1
# basic
h5py
pyyaml
tqdm
# utilties
GitPython
tabulate
gdown>=4.6.0
# vision
transforms3d
opencv-python
imageio
imageio[ffmpeg]
trimesh
rtree